Refactor download, add tests
