+++ eotdl/eotdl/access/__init__.py
@@ -1,5 +1,3 @@
-from .sentinelhub import (SHClient, 
-                          SHParameters, 
-                          EvalScripts)
-from .airbus import (AirbusClient,
-                     get_airbus_access_token)
+from .download import (download_sentinel_imagery,
+                        search_and_download_sentinel_imagery)
+from .search import search_sentinel_imagery
