+++ eotdl/eotdl/access/search.py
@@ -1,14 +1,14 @@
 from .sentinelhub import SHClient
-from .sentinelhub.parameters import (SUPPORTED_SENSORS, SH_PARAMETERS_DICT)
+from .sentinelhub.parameters import SH_PARAMETERS_DICT
+from .sentinelhub.utils import evaluate_sentinel_parameters
 from typing import Optional
 
 
 def search_sentinel_imagery(sensor: str,
-                   time_interval: Optional[list] = None,
-                   bounding_box: Optional[list] = None
-                   ) -> None:
-    if sensor not in SUPPORTED_SENSORS:
-        raise ValueError(f"Sensor {sensor} is not supported. Supported sensors are: {SUPPORTED_SENSORS}")
+                            time_interval: Optional[list] = None,
+                            bounding_box: Optional[list] = None
+                            ) -> None:
+    evaluate_sentinel_parameters(sensor, time_interval, bounding_box, output_needed=False)
 
     client = SHClient()
     parameters = SH_PARAMETERS_DICT[sensor]()
