--- eotdl/eotdl/tools/sen12floods/__init__.py
@@ -1,3 +1,0 @@
-from .tools import (calculate_average_coordinates_distance,
-                    generate_new_locations_bounding_boxes,
-                    )
